requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
brotlipy==0.7.0
certifi
cffi
charset-normalizer
colorama
contourpy==1.1.0
cryptography
cycler==0.12.1
debugpy==1.6.3
decorator==5.1.1
dill==0.3.7
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.1
fonttools==4.47.2
ibm-cloud-sdk-core==3.19.2
ibm-platform-services==0.53.2
idna
importlib-resources==6.1.1
ipykernel==6.16.0
ipython==8.13.0
jedi==0.19.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.5
matplotlib==3.7.2
matplotlib-inline==0.1.6
mido==1.3.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
ntlm-auth==1.5.0
numpy==1.23.5
packaging==23.2
parso==0.8.3
pbr==5.11.1
pillow==10.2.0
pluggy
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.2
pure-eval==0.2.2
pycparser
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
python-osc==1.8.3
python-rtmidi==1.5.7
pyzmq==24.0.1
qiskit-aer==0.13.3
qiskit-ibm-runtime==0.22.0
qiskit-ibmq-provider==0.20.2
qiskit-optimization==0.5.0
qiskit-terra==0.45.3
requests==2.31.0
requests-ntlm==1.1.0
ruamel.yaml
ruamel.yaml.clib
rustworkx==0.14.2
scipy==1.10.1
six
stack-data==0.6.3
stevedore==5.1.0
symengine==0.11.0
sympy==1.12
toolz
tornado==6.2
tqdm
traitlets==5.13.0
typing_extensions==4.8.0
urllib3==2.2.1
wcwidth==0.2.9
websocket-client==1.6.1
websockets==11.0.3
zipp==3.17.0
zstandard==0.18.0